Initial Commit
First map loader release, I've been working on this since the end of June and I just haven't had the time to fully commit.

Over the past two months I've been spending the time to learn how the custom map system works and how to load it, I've had many iterations that were unstable/crash guaranteed.

Today I finally reworked my map loader logic to work within the game thread while safely waiting for the map to finish loading, this is the first release of it but I will be reworking some things in the future, I was originally planning to make this an IMGUI implementation but unfortunately I lost motivation and mental capacity to finish it, hopefully I'll finish it later.